UI and DB update

+++ data/src/main/java/com/ethernet389/data/dao/NoteDao.kt
@@ -16,5 +16,5 @@ interface NoteDao {
     fun getAllNotes(): List<DataNote>
 
     @Delete
-    fun deleteNote(note: DataNote): Long
+    fun deleteNote(note: DataNote): Int
 }
